# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Empresa",
        "description": "Documentação da API de Empresas, Funcionários e Clientes",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista todos os clientes",
                "description": "Retorna uma lista paginada de clientes",
                "operationId": "getClientesList",
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastra um novo cliente",
                "description": "Cria um novo registro de cliente",
                "operationId": "storeCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "senha",
                                    "cpf"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Empresa X"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "cliente@example.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "documento": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "empresa_ids[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Detalhes de um cliente",
                "description": "Retorna os dados de um cliente específico",
                "operationId": "getClienteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualiza um cliente",
                "description": "Atualiza os dados de um cliente existente (Use POST com _method=PUT para upload)",
                "operationId": "updateCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Empresa X Alterada"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "cliente@example.com"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "documento": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "empresa_ids[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Remove um cliente",
                "description": "Deleta o registro de um cliente do sistema",
                "operationId": "deleteCliente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente removido com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": [
            []
        ]
    }
}